// src/app/shared/interfaces/iprofile.ts
export interface IProfile {
  fullName: string;
  email: string;
  registerType: string;
  phone: string;
  gender: string;
  dateOfBirth: Date;
  memberID: string;
  twitter: string;
  instagram: string;
  facebook: string;
  receiveClubEmails: boolean;
  shareMyEmail: boolean;
  rating: number;
}